requirements.txt
discord.py[voice, speed] ~= 2.1.0
aiofiles
python-magic-bin
yt-dlp
